requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
pydantic-settings
python-dotenv
httpx
python-docx
